requirements.txt
python-telegram-bot==21.9
python-dotenv>=1.0.0
openai>=1.12.0
pytest>=7.4.0
